add author and permission

+++ rikai/app/Http/Controllers/Admin/BookController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Book;
+use App\Models\Role;
 use App\Models\Book_category;
 use App\Models\Category;
 use Illuminate\Support\Str;
@@ -12,6 +13,9 @@ use App\Tag;
 use App\Http\Requests\BookRequest;
 use App\Library\Services\Contracts\UploadimageServiceInterface; 
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
+use App\Models\Permission;
+use App\Enums\PermissionType;
 
 class BookController extends Controller
 {
@@ -56,33 +60,44 @@ class BookController extends Controller
      */
     public function store(BookRequest $request)
     {
-        $data = $request->all();
-        $type= 'book';
-        $data['image'] = $this->uploadImageService->uploadImage($request,$data,$type);
-        $tags = explode(',',$request->tag);
-        $book = Book::create($data);
-        $book->tag($tags);
-        $categorys = $request->input('category_id');
-        foreach($categorys as $category){
-            DB::beginTransaction();
-            try{
-                $bookcategory = DB::table('book_category')->insert([
-                    'book_id' => $book->id,
-                    'category_id' => $category,
-                ]);
-                DB::commit();
-            } catch(Exception $e) {
-                DB::rollBack();
-                throw new Exception($e->getMessage());
+        $role = Role::where('id','=',Auth::user()->roles()->value('role_id'))->first();
+        $permissions = $role->permissions()->where('name','=',PermissionType::AddBook)->first();
+
+        if($permissions){
+            $data = $request->all();
+            $type= 'book';
+            $data['image'] = $this->uploadImageService->uploadImage($request,$data,$type);
+            $tags = explode(',',$request->tag);
+            $book = Book::create($data);
+            $book->tag($tags);
+            $categorys = $request->input('category_id');
+            foreach($categorys as $category){
+                DB::beginTransaction();
+                try{
+                    $bookcategory = DB::table('book_category')->insert([
+                        'book_id' => $book->id,
+                        'category_id' => $category,
+                    ]);
+                    DB::commit();
+                } catch(Exception $e) {
+                    DB::rollBack();
+                    throw new Exception($e->getMessage());
+                }
+
             }
+            if ($book){
+                $message = 'message.add_book_success';
+                return redirect()->route('bookadmin.index')->withMessage(__($message));
+            } else {
+                $message = 'message.add_book_fail';
+                return redirect()->route('bookadmin.index')->withMessage(__($message));
+            } 
+        }else{
+            $error = 'message.sufficient_permissions';
+            return redirect()->route('bookadmin.index')->withErrors(__($error));
         }
-        if ($book){
-            $message = 'message.add_book_success';
-            return redirect()->route('bookadmin.index')->withMessage(__($message));
-        } else {
-            $message = 'message.add_book_fail';
-            return redirect()->route('bookadmin.index')->withMessage(__($message));
-        }
+            
+        
     }
 
     /**
@@ -119,56 +134,63 @@ class BookController extends Controller
      */
     public function update(BookRequest $request, $bookid)
     {
-        //
-        $book = $this->findBook($bookid);
-        $data = $request->all();
-        $categorybooks = Book_Category::where('book_id','=',$bookid)->get();
-        $categorys = $request->input('category_id');
-        $categoryvalue=[];
-        foreach($categorybooks as $categorybook){
-            $categoryvalue[] = $categorybook['category_id'];
-        }
-        foreach($categorys as $category){
-            if(!in_array($category,$categoryvalue)){
-                DB::beginTransaction();
-                try{
-                    $bookcategory = DB::table('book_category')->insert([
-                        'book_id' => $book->id,
-                        'category_id' => $category,
-                    ]);
-                    DB::commit();
-                } catch(Exception $e){
-                    DB::rollBack();
-                    throw new Exception($e->getMessage());
+        $role = Role::where('id','=',Auth::user()->roles()->value('role_id'))->first();
+        $permissions = $role->permissions()->where('name','=',PermissionType::UpdateBook)->first();
+
+        if($permissions){
+            $book = $this->findBook($bookid);
+            $data = $request->all();
+            $categorybooks = Book_Category::where('book_id','=',$bookid)->get();
+            $categorys = $request->input('category_id');
+            $categoryvalue=[];
+            foreach($categorybooks as $categorybook){
+                $categoryvalue[] = $categorybook['category_id'];
+            }
+            foreach($categorys as $category){
+                if(!in_array($category,$categoryvalue)){
+                    DB::beginTransaction();
+                    try{
+                        $bookcategory = DB::table('book_category')->insert([
+                            'book_id' => $book->id,
+                            'category_id' => $category,
+                        ]);
+                        DB::commit();
+                    } catch(Exception $e){
+                        DB::rollBack();
+                        throw new Exception($e->getMessage());
+                    }
                 }
             }
-        }
-        foreach($categoryvalue as $value){
-            if(!in_array($value,$categorys)){
-                DB::beginTransaction();
-                try{
-                    $bookcategory = DB::table('book_category')->where([
-                        ['book_id' ,'=', $bookid],
-                        ['category_id', '=', $value]
-                    ])->delete();
-                    DB::commit();
-                } catch(Exception $e){
-                    DB::rollBack();
-                    throw new Exception($e->getMessage());
+            foreach($categoryvalue as $value){
+                if(!in_array($value,$categorys)){
+                    DB::beginTransaction();
+                    try{
+                        $bookcategory = DB::table('book_category')->where([
+                            ['book_id' ,'=', $bookid],
+                            ['category_id', '=', $value]
+                        ])->delete();
+                        DB::commit();
+                    } catch(Exception $e){
+                        DB::rollBack();
+                        throw new Exception($e->getMessage());
+                    }
                 }
             }
-        }
-        $type = 'book';
-        $data['image'] = $this->uploadImageService->uploadImage($request,$data,$type);
-        $tags = explode(',',$request->tag);
-        $book->update($data);
-        $book->retag($tags);
-        if($book){
-            $message = 'message.update_book_success';
-            return redirect()->route('bookadmin.edit',[$book->id])->withMessage(__($message));
-        } else {
-            $message = 'message.update_book_fail';
-            return redirect()->route('bookadmin.edit',[$book->id])->withMessage(__($message));
+            $type = 'book';
+            $data['image'] = $this->uploadImageService->uploadImage($request,$data,$type);
+            $tags = explode(',',$request->tag);
+            $book->update($data);
+            $book->retag($tags);
+            if($book){
+                $message = 'message.update_book_success';
+                return redirect()->route('bookadmin.edit',[$book->id])->withMessage(__($message));
+            } else {
+                $message = 'message.update_book_fail';
+                return redirect()->route('bookadmin.edit',[$book->id])->withMessage(__($message));
+            }
+        }else{
+            $error = 'message.sufficient_permissions';
+            return redirect()->route('bookadmin.index')->withErrors(__($error));
         }
 
     }
@@ -181,25 +203,33 @@ class BookController extends Controller
      */
     public function destroy($bookid)
     {
-        //
-        $book = $this->findBook($bookid);
-        $categories = Book_Category::where('book_id','=',$bookid)->get();
-        $book->delete();
-        foreach($categories as $category){
-            DB::beginTransaction();
-            try{
-                $bookcategory = DB::table('book_category')->where([
-                    ['book_id' ,'=', $bookid],
-                    ['category_id', '=', $category->category_id]
-                ])->delete();
-                DB::commit();
-            } catch(Exception $e){
-                DB::rollBack();
-                throw new Exception($e->getMessage());
+        $role = Role::where('id','=',Auth::user()->roles()->value('role_id'))->first();
+        $permissions = $role->permissions()->where('name','=',PermissionType::DeleteBook)->first();
+
+        if($permissions){
+            $book = $this->findBook($bookid);
+            $categories = Book_Category::where('book_id','=',$bookid)->get();
+            $book->delete();
+            foreach($categories as $category){
+                DB::beginTransaction();
+                try{
+                    $bookcategory = DB::table('book_category')->where([
+                        ['book_id' ,'=', $bookid],
+                        ['category_id', '=', $category->category_id]
+                    ])->delete();
+                    DB::commit();
+                } catch(Exception $e){
+                    DB::rollBack();
+                    throw new Exception($e->getMessage());
+                }
             }
+            $message = 'message.delete_book_success';
+            return redirect()->route('bookadmin.index')->withMessage(__($message));
+        }else{
+            $error = 'message.sufficient_permissions';
+            return redirect()->route('bookadmin.index')->withErrors(__($error));
         }
-        $message = 'message.delete_book_success';
-        return redirect()->route('bookadmin.index')->withMessage(__($message));
+
     }
 
     public function findBook($bookid){
